requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain
pandas
matplotlib
seaborn
streamlit
huggingface_hub
transformers
langchain-huggingface 